pyproject.toml
[tool.poetry]
name = "disable-graphql-introspection-django"
version = "0.0.1"
description = ""
authors = ["satoshi <nathanuchia@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.1.1"
graphene-django = "^3.2.2"
graphene-django-query-optimizer = "^0.10.1"
django-cors-headers = "^3.13.0"
django-filter = "^24.2"
gunicorn = "^22.0.0"
whitenoise = "^6.6.0"
django-allow-cidr = "^0.7.1"
pytest-django = "^4.8.0"
pytest = "^8.2.2"
django-test-plus = "^2.2.4"
pytest-sugar = "^1.0.0"
pytest-clarity = "^1.0.1"
pytest-instafail = "^0.5.0"
pytest-html = "^4.1.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
black = {version = "^22.10.0", allow-prereleases = true}
mypy = "^0.991"
ipykernel = "^6.21.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

